# src/main/resources/api.yml
openapi: 3.0.1
info:
  title: Bank Transactions API
  description: API para gestionar transacciones bancarias en el Banco XYZ
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Servidor local
paths:
  /transactions/deposit:
    post:
      summary: Registrar un depósito
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '201':
          description: Depósito registrado exitosamente
  /transactions/withdrawal:
    post:
      summary: Registrar un retiro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '201':
          description: Retiro registrado exitosamente
  /transactions/transfer:
    post:
      summary: Registrar una transferencia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '201':
          description: Transferencia registrada exitosamente
  /transactions/record:
    get:
      summary: Consultar el historial de transacciones
      responses:
        '200':
          description: Lista de transacciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
components:
  schemas:
    TransactionRequest:
      type: object
      properties:
        amount:
          type: number
          example: 100.50
        sourceAccount:
          type: string
          example: "12345678"
    TransferRequest:
      type: object
      properties:
        amount:
          type: number
          example: 100.50
        sourceAccount:
          type: string
          example: "12345678"
        destinationAccount:
          type: string
          example: "87654321"
    Transaction:
      type: object
      properties:
        id:
          type: string
          example: "64ab3c2b4d56f15c3b5e8a91"
        type:
          type: string
          example: "Deposit"
        amount:
          type: number
          example: 100.50
        date:
          type: string
          format: date-time
          example: "2024-11-15T14:12:00Z"
        sourceAccount:
          type: string
          example: "12345678"
        destinationAccount:
          type: string
          example: "87654321"